Fixing init.d's force-stop
Apparently the force-stop function of the init.d file was not working
at all here because dashes (-) are not allowed in function names.
I wonder how this was supposed to work anyway but perhaps it was broken
by an unrelated change in the shell or subleties on my system.
Moved the force-stop into the case/esac construct.